examples/aiassistant/requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
langchain>=0.1.0
python-dotenv>=0.19.0
langchain-community>=0.3.17
openai>=1.1.0